Moved utils to pkg
