+++ pkg/utils/typescanner.go
